Add files via upload
The Final Computer Assignment of this course is a simplified implementation of an OFDM system comprised of multiple modules in the transmitter and the corresponding modules in the receiver. 

I've written the report in Persian. However, plots included in the main (.mlx) file (and also the last chapter of the report), are basically answers to the following questions:

1. Plot BER over SNR for an OFDM system in the presence of an AWGN channel.

2. Plot BER over the Clipping Level of this system and deduct how the performance of the system is affected by the saturative behavior of a Power Amplifier placed right before the channel. 

3. Plot BER over SNR for a Rayleigh Fading Channel

4. Repeat Problem #3 for a situation where the receiver also includes an MMSE Equalizer. 

5. Design a Pilot Signal and leverage your system to an automatic start and end detection (using the correlation operation)